refactor: move HotelType definition to shared types for better code organization (closes #27)
