+++ backend/src/models/hotels.ts
@@ -1,21 +1,5 @@
 import mongoose from "mongoose";
-
-export type HotelType = {
-  _id: string;
-  userId: string;
-  name: string;
-  city: string;
-  country: string;
-  description: string;
-  type: string;
-  adultCount: number;
-  childCount: number;
-  facilities: string[];
-  pricePerNight: number;
-  starRating: number;
-  imageUrls: string[];
-  lastUpdated: Date;
-};
+import { HotelType } from "../shared/types";
 
 const hotelSchema = new mongoose.Schema<HotelType>({
   userId: { type: String, required: true },
