+++ frontend/src/api-client.ts
@@ -1,6 +1,6 @@
+import { HotelType } from "../../backend/src/shared/types";
 import { RegisterFormData } from "./pages/Register";
 import { SignInFormData } from "./pages/SignIn";
-import { HotelType } from "../../backend/src/models/hotels";
 
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || "";
